requirements.txt
Flask>=2.2.5
gunicorn>=20.1.0
pymongo>=4.3.3
python-dotenv>=1.0.0
Flask-Login
werkzeug
